Rename webhook handler file
