+++ lib/shopify_api.rb
@@ -16,7 +16,7 @@ require "concurrent"
 
 require_relative "shopify_api/inflector"
 require_relative "shopify_api/admin_versions"
-require_relative "shopify_api/webhooks/handler"
+require_relative "shopify_api/webhooks/webhook_handler"
 
 loader = Zeitwerk::Loader.for_gem
 loader.inflector = ShopifyAPI::Inflector.new(__FILE__)
