+++ test/test_helpers/fake_webhook_handler.rb
@@ -1,7 +1,7 @@
 # typed: false
 # frozen_string_literal: true
 
-require_relative "../../lib/shopify_api/webhooks/handler"
+require_relative "../../lib/shopify_api/webhooks/webhook_handler"
 module TestHelpers
   class FakeWebhookHandler
     include ShopifyAPI::Webhooks::WebhookHandler
